Take a snapshot of the AuroraWatch UK status
This class mimics the behaviour of the AuroraWatchUK class in the
aurorawatchuk module but its fields are evaluated just once, at the time
first requested. Thus the values it returns are snapshots of the status.
This is necessary when the information may be required multiple times as
it avoids the possibility that the status level could change between
uses. If the information is not required then no network traffic is
generated.