Add option to ignore errors arnd return as "unknown"
For printing the current status in strings it may on occassions be
better to print "Unknown" rather than raise an exception when the staus
cannot be obtained. Exceptions can only be ignored for status_level,
status_color, status_meaning and status_description; all other
information is too structured for error values to be useful.